test(notification): move tests into component folders
