+++ tests/InlineNotification/InlineNotification.test.svelte
@@ -3,14 +3,18 @@
   import type { ComponentProps } from "svelte";
 
   export let kind: ComponentProps<InlineNotification>["kind"] = "error";
-  export let lowContrast = false;
-  export let timeout = 0;
-  export let role = "alert";
-  export let title = "Error:";
-  export let subtitle = "An internal server error occurred.";
-  export let hideCloseButton = false;
-  export let statusIconDescription = "";
-  export let closeButtonDescription = "";
+  export let lowContrast: ComponentProps<InlineNotification>["lowContrast"] = false;
+  export let timeout: ComponentProps<InlineNotification>["timeout"] = 0;
+  export let role: ComponentProps<InlineNotification>["role"] = "alert";
+  export let title: ComponentProps<InlineNotification>["title"] = "";
+  export let subtitle: ComponentProps<InlineNotification>["subtitle"] = "";
+  export let statusIconDescription: ComponentProps<InlineNotification>["statusIconDescription"] =
+    undefined;
+  export let closeButtonDescription: ComponentProps<InlineNotification>["closeButtonDescription"] =
+    "Close notification";
+  export let hideCloseButton: ComponentProps<InlineNotification>["hideCloseButton"] = false;
+  export let onclose: ((event: CustomEvent) => void) | undefined = undefined;
+  export let onclick: ((event: MouseEvent) => void) | undefined = undefined;
 </script>
 
 <InlineNotification
@@ -20,10 +24,10 @@
   {role}
   {title}
   {subtitle}
-  {hideCloseButton}
   {statusIconDescription}
   {closeButtonDescription}
-  on:close={(e) => {
-    console.log("close", e.detail);
-  }}
+  {hideCloseButton}
+  on:close={onclose}
+  on:click={onclick}
+  {...$$restProps}
 />
