+++ tests/ToastNotification/ToastNotification.test.svelte
@@ -3,16 +3,20 @@
   import type { ComponentProps } from "svelte";
 
   export let kind: ComponentProps<ToastNotification>["kind"] = "error";
-  export let lowContrast = false;
-  export let timeout = 0;
-  export let role = "alert";
-  export let title = "Error";
-  export let subtitle = "An internal server error occurred.";
-  export let caption = "2024-03-21 12:00:00";
-  export let hideCloseButton = false;
-  export let fullWidth = false;
-  export let statusIconDescription = "";
-  export let closeButtonDescription = "";
+  export let lowContrast: ComponentProps<ToastNotification>["lowContrast"] = false;
+  export let timeout: ComponentProps<ToastNotification>["timeout"] = 0;
+  export let role: ComponentProps<ToastNotification>["role"] = "alert";
+  export let title: ComponentProps<ToastNotification>["title"] = "";
+  export let subtitle: ComponentProps<ToastNotification>["subtitle"] = "";
+  export let caption: ComponentProps<ToastNotification>["caption"] = "";
+  export let statusIconDescription: ComponentProps<ToastNotification>["statusIconDescription"] =
+    undefined;
+  export let closeButtonDescription: ComponentProps<ToastNotification>["closeButtonDescription"] =
+    "Close notification";
+  export let hideCloseButton: ComponentProps<ToastNotification>["hideCloseButton"] = false;
+  export let fullWidth: ComponentProps<ToastNotification>["fullWidth"] = false;
+  export let onclose: ((event: CustomEvent) => void) | undefined = undefined;
+  export let onclick: ((event: MouseEvent) => void) | undefined = undefined;
 </script>
 
 <ToastNotification
@@ -23,11 +27,11 @@
   {title}
   {subtitle}
   {caption}
-  {hideCloseButton}
-  {fullWidth}
   {statusIconDescription}
   {closeButtonDescription}
-  on:close={(e) => {
-    console.log("close", e.detail);
-  }}
+  {hideCloseButton}
+  {fullWidth}
+  on:close={onclose}
+  on:click={onclick}
+  {...$$restProps}
 />
